move load_locale codegen to it's own crate
