+++ leptos_i18n_codegen/.gitignore
@@ -1,0 +1,2 @@
+/target
+/Cargo.lock
