+++ leptos_i18n_codegen/src/load_locales/ranges.rs
