+++ leptos_i18n_codegen/src/utils/mod.rs
@@ -3,8 +3,6 @@ use quote::{format_ident, quote, ToTokens, TokenStreamExt};
 use syn::Token;
 
 pub mod formatter;
-// pub mod key;
-pub mod scoped;
 
 pub enum Keys {
     SingleKey(syn::Ident),
