+++ leptos_i18n_macro/Cargo.toml
@@ -26,28 +26,45 @@ fixed_decimal = { workspace = true, features = ["ryu"] }
 json5 = { version = "0.4", optional = true }
 leptos_i18n_parser = { workspace = true }
 tinystr = { workspace = true }
+leptos_i18n_codegen = { workspace = true }
 
 [features]
 default = ["json_files"]
-nightly = []
-suppress_key_warnings = ["leptos_i18n_parser/suppress_key_warnings"]
+nightly = ["leptos_i18n_codegen/nightly"]
+suppress_key_warnings = [
+    "leptos_i18n_parser/suppress_key_warnings",
+    "leptos_i18n_codegen/suppress_key_warnings",
+]
 json_files = ["leptos_i18n_parser/json_files"]
 yaml_files = ["dep:serde_yaml", "leptos_i18n_parser/yaml_files"]
 json5_files = ["dep:json5", "leptos_i18n_parser/json5_files"]
-interpolate_display = []
-track_locale_files = []
-islands = []
-show_keys_only = []
-dynamic_load = []
-hydrate = []
-csr = []
-ssr = []
-plurals = ["leptos_i18n_parser/plurals"]
-format_datetime = ["leptos_i18n_parser/format_datetime"]
-format_list = ["leptos_i18n_parser/format_list"]
-format_nums = ["leptos_i18n_parser/format_nums"]
-format_currency = ["leptos_i18n_parser/format_currency", "format_nums"]
-icu_compiled_data = []
+interpolate_display = ["leptos_i18n_codegen/interpolate_display"]
+track_locale_files = ["leptos_i18n_codegen/track_locale_files"]
+islands = ["leptos_i18n_codegen/islands"]
+show_keys_only = ["leptos_i18n_codegen/show_keys_only"]
+dynamic_load = ["leptos_i18n_codegen/dynamic_load"]
+hydrate = ["leptos_i18n_codegen/hydrate"]
+csr = ["leptos_i18n_codegen/csr"]
+ssr = ["leptos_i18n_codegen/ssr"]
+plurals = ["leptos_i18n_parser/plurals", "leptos_i18n_codegen/plurals"]
+format_datetime = [
+    "leptos_i18n_parser/format_datetime",
+    "leptos_i18n_codegen/format_datetime",
+]
+format_list = [
+    "leptos_i18n_parser/format_list",
+    "leptos_i18n_codegen/format_list",
+]
+format_nums = [
+    "leptos_i18n_parser/format_nums",
+    "leptos_i18n_codegen/format_nums",
+]
+format_currency = [
+    "leptos_i18n_parser/format_currency",
+    "format_nums",
+    "leptos_i18n_codegen/format_currency",
+]
+icu_compiled_data = ["leptos_i18n_codegen/icu_compiled_data"]
 
 [package.metadata.cargo-all-features]
 # cargo-all-features don't provide a way to always include one feature in a set, so CI will just do json...
