+++ leptos_i18n_macro/src/t_macro/mod.rs
@@ -4,7 +4,7 @@ use quote::{quote, ToTokens};
 use syn::parse_macro_input;
 
 use self::parsed_input::ParsedInput;
-use crate::utils::Keys;
+use leptos_i18n_codegen::utils::Keys;
 
 pub mod interpolate;
 pub mod parsed_input;
