+++ leptos_i18n_macro/src/t_plural/mod.rs
@@ -4,7 +4,7 @@ use syn::parse_macro_input;
 
 use parsed_input::ParsedInput;
 
-use crate::load_locales::plurals::PluralRuleType;
+use leptos_i18n_codegen::load_locales::plurals::PluralRuleType;
 
 pub mod parsed_input;
